the_internet_robot/requirements.txt
# Core Robot Framework
robotframework>=6.1.1,<7.0.0
robotframework-seleniumlibrary>=6.2.0,<7.0.0

# Configuration and Data
pyyaml>=6.0.1,<7.0.0
faker>=19.6.2,<20.0.0

# Reporting
allure-robotframework>=2.13.2,<3.0.0

# HTTP Requests
requests>=2.31.0,<3.0.0

# Additional utilities
selenium>=4.15.0,<5.0.0
webdriver-manager>=4.0.1,<5.0.0